+++ CHANGELOG.md
@@ -13,6 +13,7 @@ project adheres to
 
 * `value::Unit` has a new alternative.
 * The `List` alternative in `sass::Value` and `css::Value` is modified.
+* The `Use` alternative in `sass::Value` is modified.
 
 ### Improvements
 
@@ -25,6 +26,7 @@ project adheres to
 * Improved parameter checking for `hwb`, `alpha`, and `invert`
   functions in `sass:color` module.
 * The `if` function evaluates its arguments lazily.  Issue #107.
+* Modules can now be configured; `@use ... with ...`.  PR #109.
 * The `--include-path` cli argument is now named `--load-path`.
 * Update sass-spec test suite to 2021-05-12.  Also include "other"
   files (for `@use` and `@import`) in rust code for the suite.